Separate out the defaults merger from parser
