+++ components/support/nimbus-fml/src/lib.rs
@@ -2,8 +2,9 @@
 * License, v. 2.0. If a copy of the MPL was not distributed with this
 * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
+pub(crate) mod defaults_merger;
 pub mod error;
-pub mod frontend;
+pub(crate) mod frontend;
 pub mod intermediate_representation;
 pub mod parser;
 pub mod util;
